medclaim-ai/backend/requirements.txt
# Core FastAPI and web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Google ADK and AI - Production versions
google-adk>=0.1.0
google-generativeai>=0.3.2
google-cloud-aiplatform>=1.87.0

# Document processing
PyPDF2>=3.0.1
pdfplumber>=0.10.3
pytesseract>=0.3.10
Pillow>=10.1.0
python-magic>=0.4.27
pymupdf>=1.23.8

# Advanced document parsing (optional)
llama-parse>=0.4.0

# LangChain for document processing
langchain>=0.0.350
langchain-community>=0.0.10
tiktoken>=0.5.2

# File handling
aiofiles>=23.2.1

# HTTP and utilities
httpx>=0.28.1
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Additional dependencies for Google ADK
google-auth>=2.40.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.157.0